Replace static assert

+++ test/test_acosh.cpp
@@ -226,8 +226,8 @@ auto main() -> int
     local::test_acosh
     (
       static_cast<std::int32_t>(INT32_C(16) * INT32_C(262144)),
-      1.0L + static_cast<long double>(std::numeric_limits<boost::decimal::decimal32_t>::epsilon()) * 10.0L,
-      1.0L + static_cast<long double>(std::numeric_limits<boost::decimal::decimal32_t>::epsilon()) * 100.0L
+      1.0L + static_cast<double>(std::numeric_limits<boost::decimal::decimal32_t>::epsilon()) * 10.0L,
+      1.0L + static_cast<double>(std::numeric_limits<boost::decimal::decimal32_t>::epsilon()) * 100.0L
     );
 
   const auto result_tiny_is_ok   = local::test_acosh(static_cast<std::int32_t>(INT32_C(4096)), 1.001L, 1.1L);
